requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0

# Browser automation for Anubis bypass (choose one)
# Playwright (recommended)
playwright>=1.40.0

# OR Selenium (alternative)
# selenium>=4.15.0

# Optional: For better HTML parsing
lxml>=4.9.0
